Apply 1 suggestion(s) to 1 file(s)

+++ inc/Definitions.hpp
@@ -12,7 +12,7 @@ inline uint8_t const NumOfActuators = 2;
 inline uint8_t const LocalStateSize = 6;
 inline uint8_t const GlobalStateSize = 7;
 inline uint8_t const MeasurementSize = 6;
-inline int16_t const ReflectivityDataRows = 180;
+inline uint8_t const ReflectivityDataRows = 180;
 inline int16_t const ReflectivityDataColumns = 288;
 
 typedef Eigen::Matrix<float, ReflectivityDataRows, ReflectivityDataColumns> EarthCellsMatrix;
